+++ R/pep725_demo.R
@@ -33,7 +33,7 @@
 #' }
 #'
 #' @examples
-#' \dontrun{
+#' if (interactive()) {
 #' # Run all demos interactively
 #' pep725_demo()
 #'