+++ R/pheno_leaflet.R
@@ -20,8 +20,7 @@
 #' }
 #'
 #' @examples
-#' \dontrun{
-#' # Download synthetic data first
+#' if (interactive()) {
 #' pep <- pep_download()
 #'
 #' # For faster loading, filter the data first (recommended)